refactor: extract symbol ID generation into dedicated module
